Rename rule to "ShouldHaveOnlyOnePublicMethodNamed".
