--- src/Rule/Assertion/Declaration/ShouldHaveOnlyPublicMethodNamed/HasOnlyPublicMethodNamedRule.php
@@ -1,10 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace PHPat\Rule\Assertion\Declaration\ShouldHaveOnlyPublicMethodNamed;
-
-use PHPat\Rule\Extractor\Declaration\PublicMethodNamedExtractor;
-
-final class HasOnlyPublicMethodNamedRule extends ShouldHaveOnlyPublicMethodNamed
-{
-    use PublicMethodNamedExtractor;
-}
