+++ tests/unit/rules/ShouldHaveOnlyOnePublicMethodNamed/GoodImplementationClassWithOnlyOnePublicMethodNamedWithRegexTest.php
@@ -1,17 +1,16 @@
 <?php declare(strict_types=1);
 
-namespace Tests\PHPat\unit\rules\ShouldHaveOnlyPublicMethodNamed;
+namespace Tests\PHPat\unit\rules\ShouldHaveOnlyOnePublicMethodNamed;
 
 use PHPat\Configuration;
-use PHPat\Rule\Assertion\Declaration\ShouldHaveOnlyPublicMethodNamed\HasOnlyPublicMethodNamedRule;
-use PHPat\Rule\Assertion\Declaration\ShouldHaveOnlyPublicMethodNamed\HasOnlyPublicPublicMethodNamedRule;
-use PHPat\Rule\Assertion\Declaration\ShouldHaveOnlyPublicMethodNamed\ShouldHaveOnlyPublicMethodNamed;
+use PHPat\Rule\Assertion\Declaration\ShouldHaveOnlyOnePublicMethodNamed\HasOnlyOnePublicMethodNamedRule;
+use PHPat\Rule\Assertion\Declaration\ShouldHaveOnlyOnePublicMethodNamed\ShouldHaveOnlyOnePublicMethodNamed;
 use PHPat\Selector\Classname;
 use PHPat\Statement\Builder\StatementBuilderFactory;
 use PHPStan\Rules\Rule;
 use PHPStan\Testing\RuleTestCase;
 use PHPStan\Type\FileTypeMapper;
-use Tests\PHPat\fixtures\Special\ClassWithOnlyPublicMethodNamed;
+use Tests\PHPat\fixtures\Special\ClassWithOnlyOnePublicMethodNamed;
 use Tests\PHPat\unit\FakeTestParser;
 
 /**
@@ -19,27 +18,27 @@ use Tests\PHPat\unit\FakeTestParser;
  * @internal
  * @coversNothing
  */
-class GoodImplementationClassWithOnlyPublicMethodNamedWithRegexTest extends RuleTestCase
+class GoodImplementationClassWithOnlyOnePublicMethodNamedWithRegexTest extends RuleTestCase
 {
-    public const RULE_NAME = 'testFixtureClassShouldHaveOnlyPublicMethodNamed';
+    public const RULE_NAME = 'testFixtureClassShouldHaveOnlyOnePublicMethodNamed';
 
     public function testRule(): void
     {
-        $this->analyse(['tests/fixtures/Special/ClassWithOnlyPublicMethodNamed.php'], []);
+        $this->analyse(['tests/fixtures/Special/ClassWithOnlyOnePublicMethodNamed.php'], []);
     }
 
     protected function getRule(): Rule
     {
         $testParser = FakeTestParser::create(
             self::RULE_NAME,
-            ShouldHaveOnlyPublicMethodNamed::class,
-            [new Classname(ClassWithOnlyPublicMethodNamed::class, false)],
+            ShouldHaveOnlyOnePublicMethodNamed::class,
+            [new Classname(ClassWithOnlyOnePublicMethodNamed::class, false)],
             [],
             [],
             ['name' => '/^method[a-zA-Z0-9]+/', 'isRegex' => true]
         );
 
-        return new HasOnlyPublicMethodNamedRule(
+        return new HasOnlyOnePublicMethodNamedRule(
             new StatementBuilderFactory($testParser),
             new Configuration(false, true, false),
             self::createReflectionProvider(),
